Moved tests under tests folder
Although is a bit uglier (question of taste will ya):

* It separates source code form tests
* Allow a better autoloading

And doesn't change the bundle meta structure ;)
