+++ behat.yml
@@ -11,21 +11,21 @@ default:
             contexts:
                 - Behat\MinkExtension\Context\MinkContext
                 - Fidry\AliceBundleExtension\Context\Doctrine\AliceORMContext
-                - Incipio\Tests\Behat\Context\ApiContext:
+                - Incipio\Test\Behat\Context\ApiContext:
                       doctrine:       @doctrine
                       jwtManager:     @lexik_jwt_authentication.jwt_manager
                       userManager:    @fos_user.user_manager
                       encoderFactory: @security.encoder_factory
-                - Incipio\Tests\Behat\Context\JsonContext:
+                - Incipio\Test\Behat\Context\JsonContext:
                       evaluationMode: javascript
-                - Incipio\Tests\Behat\Context\JsonLdContext:
+                - Incipio\Test\Behat\Context\JsonLdContext:
                       evaluationMode: javascript
 
         front_features:
             paths:    [ %paths.base%/features/front ]
             contexts:
                 - Fidry\AliceBundleExtension\Context\Doctrine\AliceORMContext
-                - Incipio\Tests\Behat\Context\FrontContext
+                - Incipio\Test\Behat\Context\FrontContext
 
     extensions:
         Behat\MinkExtension:
