+++ composer.json
@@ -4,10 +4,16 @@
     "license": "GNU",
     "type": "project",
     "autoload": {
-        "psr-4": { "": "src/", "Incipio\\": "app/" }
+        "psr-4": {
+            "": "src/",
+            "Incipio\\": "app/"
+        }
     },
     "autoload-dev": {
-        "psr-4": { "Incipio\\Tests\\Behat\\": "features/bootstrap" }
+        "psr-4": {
+            "": "tests/",
+            "Incipio\\Test\\Behat\\": "features/bootstrap"
+        }
     },
     "require": {
         "php": ">=5.5.0",
