--- features/bootstrap/Tests/Fixtures/json-context-1.json
@@ -1,3 +1,0 @@
-{
-  "intValue": 0
-}
