--- features/bootstrap/behat.yml
@@ -1,33 +1,0 @@
-default:
-    calls:
-        error_reporting: 16383 # E_ALL & ~E_USER_DREPRECATED
-
-    formatters:
-        pretty: true
-
-    suites:
-        api_features:
-            paths:    [ %paths.base%/features/api ]
-            contexts:
-                - Incipio\Tests\Behat\Context\ApiContext:
-                      doctrine:       @doctrine
-                      jwtManager:     @lexik_jwt_authentication.jwt_manager
-                      userManager:    @fos_user.user_manager
-                      encoderFactory: @security.encoder_factory
-                - Behat\MinkExtension\Context\MinkContext
-                - Incipio\Tests\Behat\Context\JsonContext:
-                      evaluationMode: javascript
-                - Incipio\Tests\Behat\Context\JsonLdContext:
-                      evaluationMode: javascript
-
-        front_features:
-            paths:    [ %paths.base%/features/front ]
-            contexts: [ Incipio\Tests\Behat\Context\FrontContext ]
-
-    extensions:
-        Behat\MinkExtension:
-            base_url:   http://localhost:8080
-            files_path: bootstrap/Tests/Fixtures
-            goutte:    ~
-            selenium2: ~
-        Sanpi\Behatch\Extension: ~
