+++ tests/ApiBundle/Tests/DependencyInjection/ApiExtensionTest.php
