+++ tests/ApiBundle/Tests/Doctrine/ORM/Filter/UserTypeFilterTest.php
