+++ tests/ApiBundle/Tests/Doctrine/ORM/Manager/MandateNonPersistentManagerTest.php
