+++ tests/ApiBundle/Tests/Doctrine/ORM/Manager/UserManagerTest.php
