+++ tests/ApiBundle/Tests/Entity/MandateTest.php
@@ -13,6 +13,7 @@ namespace ApiBundle\Tests\Entity;
 
 use ApiBundle\Entity\Job;
 use ApiBundle\Entity\Mandate;
+use ApiBundle\Test\Entity\AbstractEntityTestCase;
 
 /**
  * @coversDefaultClass ApiBundle\Entity\Mandate
