+++ tests/ApiBundle/Tests/Entity/UserTest.php
@@ -13,6 +13,7 @@ namespace ApiBundle\Tests\Entity;
 
 use ApiBundle\Entity\Job;
 use ApiBundle\Entity\User;
+use ApiBundle\Test\Entity\AbstractEntityTestCase;
 
 /**
  * @coversDefaultClass ApiBundle\Entity\User
