+++ tests/ApiBundle/Tests/Utils/UserRolesTest.php
