+++ tests/FrontBundle/Tests/Client/ApiClientTest.php
