+++ tests/FrontBundle/Tests/Controller/BaseControllerTest.php
