+++ tests/FrontBundle/Tests/DependencyInjection/FrontExtensionTest.php
