+++ tests/FrontBundle/Tests/Form/Type/UserTypeTest.php
