+++ tests/FrontBundle/Tests/Security/Http/Authentication/AuthenticationSuccessHandlerTest.php
