+++ tests/FrontBundle/Tests/Twig/FrontExtensionTest.php
