+++ tests/FrontBundle/Tests/Utils/IriHelperTest.php
