+++ tests/FrontBundle/Tests/Utils/RoleHierarchyHelperTest.php
